fix(plugins): remove legacypaths load-time self-healing
The internal/legacypaths package rewrote Serf→Evener path prefixes at
load time in the plugins package (rerootLegacyInstallLocations,
rerootLegacyInstallPaths, legacyRootFor). This was an open-ended
back-compat commitment: every load silently healed stale paths from
the pre-rename era without requiring the user to run evener-migrate.

In zero-back-compat land, load-time self-healing should not exist. The
migration tool (cmd/evener-migrate) is the proper place for one-time
path rewriting; users who haven't migrated should run it once.

Removed:
- internal/legacypaths package (legacypaths.go, legacypaths_test.go)
- rerootLegacyInstallLocations from marketplaces.go + its call
- rerootLegacyInstallPaths from registry.go + its call
- legacyRootFor from paths.go
- 4 legacy tests (TestLegacyRootFor, TestLoadRegistry_RerootsLegacyInstallPath,
  TestLoadMarketplaces_RerootsLegacyInstallLocation,
  TestRefreshMarketplace_RepairsLegacyInstallLocation)

The Rewrite function is inlined into cmd/evener-migrate/rewrite.go as
rewritePathPrefix (the migrate tool is the legitimate one-time user).
Its 9 unit test cases are ported to rewrite_prefix_test.go.

simplify-code: 4 reviewers (reuse/simplification/efficiency/altitude all
clean; reuse flagged test-coverage gap from inlining — fixed by porting
tests; altitude flagged gofmt nit — fixed).
