+++ cmd/evener-migrate/rewrite.go
@@ -7,9 +7,8 @@ import (
 	"io/fs"
 	"os"
 	"path/filepath"
+	"strings"
 	"unicode/utf8"
-
-	"primeradiant.com/evener/internal/legacypaths"
 )
 
 // maxRewriteFileSize caps how large a file rewriteLegacyPaths will read into
@@ -77,7 +76,7 @@ func rewriteLegacyPaths(dst, oldRoot, newRoot string, stdout io.Writer) error {
 			return nil
 		}
 
-		rewritten, n := legacypaths.Rewrite(string(data), oldRoot, newRoot)
+		rewritten, n := rewritePathPrefix(string(data), oldRoot, newRoot)
 		if n == 0 {
 			return nil
 		}
@@ -98,4 +97,47 @@ func looksBinary(data []byte) bool {
 		sniff = sniff[:binarySniffWindow]
 	}
 	return bytes.Contains(sniff, []byte{0}) || !utf8.Valid(data)
+}
+
+// rewritePathPrefix replaces every occurrence of oldRoot in content with
+// newRoot, but only when oldRoot is immediately followed by a path separator
+// or any other byte that could not extend the same path component (e.g. the
+// closing quote of a JSON string, or the end of content). This prevents a
+// false match against an unrelated, longer path component — oldRoot
+// "/a/.serf" must not match inside "/a/.serfbackup".
+//
+// It reports how many replacements were made. An empty oldRoot, or
+// oldRoot == newRoot, is a no-op.
+func rewritePathPrefix(content, oldRoot, newRoot string) (string, int) {
+	if oldRoot == "" || oldRoot == newRoot {
+		return content, 0
+	}
+
+	var b strings.Builder
+	n := 0
+	rest := content
+	for {
+		i := strings.Index(rest, oldRoot)
+		if i < 0 {
+			b.WriteString(rest)
+			break
+		}
+		end := i + len(oldRoot)
+		b.WriteString(rest[:i])
+		if end == len(rest) || !isPathContinuation(rest[end]) {
+			b.WriteString(newRoot)
+			n++
+		} else {
+			b.WriteString(rest[i:end])
+		}
+		rest = rest[end:]
+	}
+	return b.String(), n
+}
+
+// isPathContinuation reports whether b could be part of the same path
+// component as the byte before it (i.e. it does not mark a path boundary).
+func isPathContinuation(b byte) bool {
+	return b == '_' || b == '-' || b == '.' ||
+		(b >= 'a' && b <= 'z') || (b >= 'A' && b <= 'Z') || (b >= '0' && b <= '9')
 }
